requirements.txt
crewai
python-dotenv
langchain-google-genai
PyGithub
Flask